+++ package.json
@@ -1,14 +1,21 @@
 {
   "name": "under-new-management",
-  "version": "0.0.2",
-  "description": "Early warning system for acquired extensions",
+  "description": "early warning system for acquired extensions",
+  "version": "0.1.0",
+  "type": "module",
   "scripts": {
-    "start": "parcel watch src/manifest.json --host localhost --config @parcel/config-webextension",
-    "build": "parcel build src/manifest.json --config @parcel/config-webextension"
+    "dev": "wxt",
+    "dev:firefox": "wxt -b firefox",
+    "build": "wxt build",
+    "build:firefox": "wxt build -b firefox",
+    "zip": "wxt zip",
+    "zip:firefox": "wxt zip -b firefox",
+    "compile": "tsc --noEmit",
+    "postinstall": "wxt prepare"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/msfrisbie/under-new-management.git"
+    "url": "git+https://github.com/maxtheaxe/under-new-management-firefox.git"
   },
   "author": "max@tnkr.xyz",
   "license": "MIT",
@@ -16,33 +23,22 @@
     "url": "https://github.com/maxtheaxe/under-new-management-firefox/issues"
   },
   "homepage": "https://github.com/maxtheaxe/under-new-management-firefox#readme",
-  "devDependencies": {
-    "@parcel/config-webextension": "^2.12.0",
-    "assert": "^2.1.0",
-    "browserify-zlib": "^0.2.0",
-    "buffer": "^6.0.3",
-    "crypto-browserify": "^3.12.0",
-    "events": "^3.3.0",
-    "https-browserify": "^1.0.0",
-    "os-browserify": "^0.3.0",
-    "parcel": "^2.12.0",
-    "path-browserify": "^1.0.1",
-    "postcss": "^8.4.35",
-    "process": "^0.11.10",
-    "punycode": "^2.3.1",
-    "stream-browserify": "^3.0.0",
-    "stream-http": "^3.2.0",
-    "string_decoder": "^1.3.0",
-    "tailwindcss": "^3.4.1",
-    "url": "^0.11.3",
-    "util": "^0.12.5",
-    "vm-browserify": "^1.1.2"
-  },
   "dependencies": {
-    "@types/chrome": "^0.0.263",
-    "@types/firefox-webext-browser": "^120.0.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "webextension-polyfill": "^0.12.0"
+  },
+  "devDependencies": {
+    "@types/lodash-es": "^4.17.12",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
+    "@types/webextension-polyfill": "^0.12.1",
+    "@wxt-dev/module-react": "^1.1.1",
+    "autoprefixer": "^10.4.20",
     "lodash-es": "^4.17.21",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "postcss": "^8.4.49",
+    "tailwindcss": "^3.4.15",
+    "typescript": "^5.6.3",
+    "wxt": "^0.19.13"
   }
 }
