--- poscss.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  plugins: [require("tailwindcss"), require("autoprefixer")],
-};
