+++ postcss.config.js
@@ -1,0 +1,10 @@
+// postcss.config.js
+import tailwindcss from 'tailwindcss';
+import autoprefixer from 'autoprefixer';
+
+export default {
+  plugins: {
+    tailwindcss,
+    autoprefixer,
+  },
+};
