+++ src/entrypoints/popup/index.html
@@ -6,6 +6,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="./index.ts"></script>
+    <script type="module" src="main.ts"></script>
   </body>
 </html>
