+++ src/entrypoints/popup/style.css
