--- src/manifest.json
@@ -1,18 +1,0 @@
-{
-  "name": "Under New Management",
-  "version": "0.0.2",
-  "description": "Detect when your installed browser extensions have changed owners.",
-  "manifest_version": 2,
-  "background": {
-    "scripts": ["background.ts"]
-  },
-  "browser_action": {
-    "default_popup": "popup/index.html"
-  },
-  "permissions": ["alarms", "management", "storage"],
-  "icons": {
-    "16": "logo16.png",
-    "48": "logo48.png",
-    "128": "logo128.png"
-  }
-}
